+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.1.1 September 8, 2020
+* Added body class bcgov-menu-primary-is-desktop to notify other components not with in menu.
+* Add menu close if clicked outside of menu (mobile)
+* Add close on mobile menu
+* fix menu arrows.
+
 ### 1.1.0 August 31, 2020
 * Changed css for menu
 * Added additional css variables, to make overrides easier.